Changes to how the API and Click do recv(from).
- When a stream  application asks for less data than is in the packet, the remainder is
now stored in click rather than in the socket state in the API.
- If a udp applications asks for less data than is in the packet, the remainder is now
discarded rather than saved.
- MSG_PEEK is now supported and is handled inside of click.

- Also changed syn handling to return POLLIN as well as POLLOUT when a SYN is received.

+++ click-2.0.1/elements/xia/xtransport.cc
@@ -624,12 +624,20 @@ int XTRANSPORT::read_from_recv_buf(xia::XSocketMsg *xia_socket_msg, sock *sk) {
 
 	if (sk->sock_type == SOCK_STREAM) {
 //		printf("<<< read_from_recv_buf: port=%u, recv_base=%d, next_recv_seqnum=%d, recv_buf_size=%d\n", sk->port, sk->recv_base, sk->next_recv_seqnum, sk->recv_buffer_size);
+
 		xia::X_Recv_Msg *x_recv_msg = xia_socket_msg->mutable_x_recv();
 		int bytes_requested = x_recv_msg->bytes_requested();
+		bool peek = x_recv_msg->peek();
 		int bytes_returned = 0;
-		char buf[1024*1024]; // TODO: pick a buf size
-		memset(buf, 0, 1024*1024);
-		for (int i = sk->recv_base; i < sk->next_recv_seqnum; i++) {
+
+		// FIXME - this should use the recv buffer size
+		char buf[64*1024]; // TODO: pick a buf size
+		memset(buf, 0, 64*1024);
+		int i;
+
+		// FIXME: make sure bytes requested is <= recv buffer size
+
+		for (i = sk->recv_base; i < sk->next_recv_seqnum; i++) {
 
 			if (bytes_returned >= bytes_requested) break;
 
@@ -638,16 +646,46 @@ int XTRANSPORT::read_from_recv_buf(xia::XSocketMsg *xia_socket_msg, sock *sk) {
 			TransportHeader thdr(p);
 			size_t data_size = xiah.plen() - thdr.hlen();
 
-			memcpy((void*)(&buf[bytes_returned]), (const void*)thdr.payload(), data_size);
+			const char *payload = (char *)thdr.payload();
+			uint16_t tail = XIA_TAIL_ANNO(p);
+
+			if (tail) {
+				click_chatter("packet (%d) has %d bytes of %d remaining\n", i % sk->recv_buffer_size, data_size - tail, data_size);
+				data_size -= tail;
+				payload += tail;
+			}
+
+			memcpy((void*)(&buf[bytes_returned]), (const void *)payload, data_size);
 			bytes_returned += data_size;
 
-			p->kill();
-			sk->recv_buffer[i % sk->recv_buffer_size] = NULL;
-			sk->recv_base++;
+			// leave the data if the user peeked
+			if (!peek) {
+				if (bytes_returned <= bytes_requested) {
+					// it's safe to delete this packet
+//					click_chatter("deleting packet %d\n", i % sk->recv_buffer_size);
+					p->kill();
+					sk->recv_buffer[i % sk->recv_buffer_size] = NULL;
+					sk->recv_base++;
+
+				} else {
+					// we need to keep the tail data the application didn't ask for
+					// update this packet to shrink the data
+					int extra = bytes_returned - bytes_requested;
+					tail = xiah.plen() - thdr.hlen() - extra;
+
+					click_chatter("keeping the last %d bytes in packet %d\n", extra, i % sk->recv_buffer_size);
+					SET_XIA_TAIL_ANNO(p, tail);
+				}
+			} else {
+				click_chatter("peeking, so leaving all data behind for packet %d\n", i % sk->recv_buffer_size);
+			}
 //			printf("    port %u grabbing index %d, seqnum %d\n", sk->port, i%sk->recv_buffer_size, i);
 		}
-		x_recv_msg->set_payload(buf, bytes_returned); // TODO: check this: need to turn buf into String first?
+
+		x_recv_msg->set_payload(buf, bytes_returned);
 		x_recv_msg->set_bytes_returned(bytes_returned);
+
+		click_chatter("returning %d bytes out of %d requested\n", bytes_returned, bytes_requested);
 
 //		printf(">>> read_from_recv_buf: port=%u, recv_base=%d, next_recv_seqnum=%d, recv_buf_size=%d\n", sk->port, sk->recv_base, sk->next_recv_seqnum, sk->recv_buffer_size);
 		return bytes_returned;
@@ -655,6 +693,8 @@ int XTRANSPORT::read_from_recv_buf(xia::XSocketMsg *xia_socket_msg, sock *sk) {
 	} else if (sk->sock_type == SOCK_DGRAM || sk->sock_type == SOCK_RAW) {
 		xia::X_Recvfrom_Msg *x_recvfrom_msg = xia_socket_msg->mutable_x_recvfrom();
 	
+		bool peek = x_recvfrom_msg->peek();
+
 		// Get just the next packet in the recv buffer (we don't return data from more
 		// than one packet in case the packets came from different senders). If no
 		// packet is available, we indicate to the app that we returned 0 bytes.
@@ -701,11 +741,18 @@ int XTRANSPORT::read_from_recv_buf(xia::XSocketMsg *xia_socket_msg, sock *sk) {
 			x_recvfrom_msg->set_sender_dag(src_path.c_str());
 			x_recvfrom_msg->set_bytes_returned(data_size);
 
-			p->kill();
-			sk->recv_buffer[sk->dgram_buffer_start] = NULL;
-			sk->recv_buffer_count--;
-			sk->dgram_buffer_start = (sk->dgram_buffer_start + 1) % sk->recv_buffer_size;
+			if (!peek) {
+				// NOTE: bytes beyond what the app asked for will be discarded, 
+				// they are not saved for the next recv like streaming socket data
+
+				p->kill();
+				sk->recv_buffer[sk->dgram_buffer_start] = NULL;
+				sk->recv_buffer_count--;
+				sk->dgram_buffer_start = (sk->dgram_buffer_start + 1) % sk->recv_buffer_size;
+			}
+
 			return data_size;
+
 		} else {
 			x_recvfrom_msg->set_bytes_returned(0);
 			return 0;
@@ -1013,7 +1060,8 @@ void XTRANSPORT::ProcessNetworkPacket(WritablePacket *p_in)
 
 				if (sk->polling) {
 					// tell API we are writeable
-					ProcessPollEvent(_dport, POLLOUT);
+					ProcessPollEvent(_dport, POLLIN|POLLOUT);
+					click_chatter("sending pollout after syn received\n");
 				}
 
 			// Mark these src & dst XID pair
@@ -3041,7 +3089,7 @@ void XTRANSPORT::Xsend(unsigned short _sport, xia::XSocketMsg *xia_socket_msg, W
 	xia::X_Send_Msg *x_send_msg = xia_socket_msg->mutable_x_send();
 	int pktPayloadSize = x_send_msg->payload().size();
 
-	char payload[16384];
+	char payload[65536];
 	memcpy(payload, x_send_msg->payload().c_str(), pktPayloadSize);
 	//click_chatter("XSEND: %d bytes from (%d)\n", pktPayloadSize, _sport);
 
@@ -3340,6 +3388,8 @@ void XTRANSPORT::Xrecv(unsigned short _sport, xia::XSocketMsg *xia_socket_msg)
 	if(sk->port != _sport) {
 		click_chatter("Xtransport::Xrecv: ERROR sk->port %d _sport %d", sk->port, _sport);
 	}
+
+	click_chatter("%d Blocking = %d\n", _sport, xia_socket_msg->blocking());
 	read_from_recv_buf(xia_socket_msg, sk);
 
 	if (xia_socket_msg->x_recv().bytes_returned() > 0) {
